made new fragment, changed code as requested
added butterknife, moved everything in main fragment to speech fragment,
made new xml for speech fragment.

*note that content_main.xml points to speech fragment, not main fragment
