+++ app/src/main/res/layout/content_main.xml
@@ -2,7 +2,7 @@
           xmlns:tools="http://schemas.android.com/tools"
           xmlns:app="http://schemas.android.com/apk/res-auto"
           android:id="@+id/fragment"
-          android:name="com.mobileproto.hireddit.hireddit.MainFragment"
+          android:name="com.mobileproto.hireddit.hireddit.SpeechFragment"
           android:layout_width="match_parent"
           android:layout_height="match_parent"
           app:layout_behavior="@string/appbar_scrolling_view_behavior"
