+++ app/src/main/res/layout/fragment_main.xml
@@ -13,10 +13,5 @@
         android:text="Hello World!"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"/>
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/speech"
-        android:text="@string/speech_button"/>
 
 </RelativeLayout>
